+++ ios/flutter_pty2/Sources/flutter_pty2/flutter_pty.c
@@ -1,0 +1,1 @@
+#include "../../../../src/flutter_pty.c"